fix: correct stock percentage color and date search
- Use parseFloat instead of parseInt for percentage comparison,
  fixing small fractional increases (e.g. 0.54%) showing green
- Return date URL from getStockUrl() so date search actually
  hits the historical data endpoint
- Strip hyphens in getConversionDate() to support YYYY-MM-DD format

+++ src/utils/stock.ts
@@ -14,7 +14,7 @@ export function getStockUpsAndDownsPercentage(
 
   return getDisplayActionText(
     percentageHandle(dec),
-    parseInt(dec) > 0 ? Status.failed : Status.success
+    parseFloat(dec) > 0 ? Status.failed : Status.success
   )
 }
